Integrate below trackers: EDK604	SupportPalette() algorithm incorrect EDK584	Remove IDT in LongMode.asm to reduce size of DxeIplX64 EDK589	GCD calculate the end of gMemoryTypeInformation error EDK590	IdeBus call ReportStatusCodeWithDevicePath with NULL path EDK587	Link one compression method in DxeIplX64 to save size EDK588	The enhancement of BdsLibUpdateConsoleVariable( ). EDK591	UOL should provide more INTxx information EDK592	UOL x64 size can be enlarged EDK594	UoL needs to detailed boot flow added to Readme.txt EDK593	UOL need more debug info for INT13 disk error EDK610	Unplug 2 more hub will cause system hang on efi32/uefi32 EDK595	Genbootsector for UoL needs enhancement EDK596	Need a special MonotonicCounter driver for UoL EDK597	UOL, BdsPlatform need check PCI_CLASS_ISA_POSITIVE_DECODE EDK600	UoL cannot boot from USB disk on some Legacy BIOS EDK599	FWimage should not assume RVA = FileOffset EDK603	UOL, dumpbs LBAOffset error EDK607	UOL, need add UART2 as default conout device in platform bds EDK608	Add IPF build support from WINDDK3790.1830 EDK609	Add DevicePath driver for uefi tip EDK226	NULL Pointer Error in ConsoleOption EDK619	READ_FORMAT_CAPACITIES command returns wrong value for USB2.0 device EDK625	Improve performance of BdsLibEnumerateAllBootOption() EDK227	NULL Pointer Error in BootMaint EDK242	NULL Pointer Errors in Variable.c EDK612	Use macro to switch the usage of PeiReportStatusCode EDK580	TCP4 Service binding: EFI_UNSUPPORTED will be returned when TCP protocol removed EDK586	Wrap assembly function with C for CpuIA32Lib, EfiCommonLib to save PEI size EDK618	CpuIo.c, Address is casted before the parameter check EDK617	UoL needs to update MBR code for some Legacy BIOS EDK611	Need to add the definition of PXE_OPFLAGS_STATION_ADDRESS_WRITE to EfiPxe.h EDK615	UOL add INT10 video support EDK621	UOL, 8254Timer should not use hardcode 0x408 EDK623	RT driver cannot output debug messages to serials in Dxe period under UoL uefi64 EDK624	X64, EFI_BREAKPOINT() not work
git-svn-id: https://svn.code.sf.net/p/efidevkit/code/trunk/Edk@181 d4c01ab0-1994-4ed3-9bc5-3ada14f9202f

+++ Sample/Platform/UoL/Loader/DumpBs/DumpBs.c
@@ -472,7 +472,7 @@ ShowBootSectorLbaOffset (
     if (PartitionIndex >= 4) {
       Print (L"dumpbs: ParitionIndex error - %d\n", PartitionIndex);
     } else {
-      LBAOffset = *(UINT64 *)((UINT8 *)Buffer + 
+      LBAOffset = (UINT64)*(UINT32 *)((UINT8 *)Buffer + 
                               MBR_START_OFFSET + 
                               MBR_ENTRY_LEN * PartitionIndex +
                               MBR_STARTINGLBA_OFFSET
